Make ProxibaseClient a singleton. Streamline signin/signout API.
Make the Proxibase client a singleton.
Isolate handling of user and session defaults to the Proxibase Client.
Have Proxibase handle the installId so callers don’t have to make
something up.
Simplify completion block signatures to remove the userid and session
parameters, which can now be retrieved from the Proxibase instance if
needed.
Add a rudimentary error string to the login screen to communicate login
errors.