Add patch muting, caching improvements, bug fixes
Users can control whether notifications for a patch are muted or active. Lots of improvements
in the networking and data layers including cancelling entity inserts/updates that are taking
too long because of poor network conditions. All S3 functionality is now consolidated in the
AwsS3 class. Improved performance for table view loading and scrolling.